handson/dojo/dynamodb/requirements.txt
aws-cdk.aws-dynamodb==1.100.0
boto3
jupyter
